Move ImageQualityGuide component over to Uploader
