+++ src/components/Capture/style.css
@@ -22,64 +22,3 @@
     background-color: @selectors-box-shadow-active;
   }
 }
-
-.imageQualityGuide {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
-.imageQualityGuideRow {
-  display: flex;
-  justify-content: space-between;
-  flex-grow: 1;
-}
-
-.documentExampleCol {
-  width: 50%;
-  font-size: 14*@unit;
-  margin-bottom: 10*@unit;
-}
-
-.documentExampleImg {
-  background-position: 50% 50%;
-  background-repeat: no-repeat;
-  background-size: contain;
-  flex-shrink: 0;
-  max-width: 100%;
-  width: 149*@unit;
-  height: 106*@unit;
-  margin: 0 auto;
-}
-
-.documentExampleImgBlur {
-  background-image: url('assets/img-blur.svg');
-}
-
-.documentExampleImgCutoff {
-  background-image: url('assets/img-cutoff.svg');
-}
-
-.documentExampleImgGlare {
-  background-image: url('assets/img-glare.svg');
-}
-
-.documentExampleImgGood {
-  background-image: url('assets/img-good.svg');
-}
-
-.documentExampleLabel {
-  text-align: center;
-}
-
-.buttonContainer {
-  @media (--small-viewport) {
-    display: flex;
-    width: 100%;
-  }
-}
-
-.error {
-  &:extend(.text);
-  color: @color-error;
-}
