+++ src/components/Uploader/Error.js
@@ -1,0 +1,13 @@
+import { h } from 'preact'
+
+import style from './style.css'
+import errors from '../strings/errors'
+
+const UploadError = ({ error, translate }) => {
+  const { message, instruction } = errors[error.name]
+  return (
+    <div className={style.error}>{`${translate(message)} ${translate(instruction)}`}</div>
+  )
+}
+
+export default UploadError
